LayoutTests/security/clip-invalid-rect-expected.txt

ALERT: Test passed because nothing crashed
